fix: query client needs to be created per request
